requirements.txt
gradio>=4.44
openai>=1.40
whisperx @ git+https://github.com/m-bain/whisperX.git
pyannote.audio>=3.1
transformers>=4.44
torch>=2.1
sentencepiece
datasets
jiwer
rouge-score
bert-score
sacrebleu
pydantic>=2.0
pandas
numpy
matplotlib
plotly
scikit-learn
faiss-cpu
diskcache
python-dotenv
pyyaml
tqdm
